requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==5.0.2
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==0.21.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
Werkzeug==3.0.1
